% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{missing_dependencies}
\alias{missing_dependencies}
\title{Get missing dependencies for list of reports.}
\usage{
missing_dependencies(report_names, dependencies)
}
\arguments{
\item{dependencies}{Names list of report to its dependencies}

\item{path}{Path to orderly root.}

\item{reports}{Report names to get missing dependencies for}
}
\value{
List with key report name and value is any missing dependencies.
}
\description{
Report A has a missing dependency if, report A has a "depends" block in
its `orderly.yml` which depends on latest version (or latest with some
search term) of report B and report B does not occur in the list of reports.
}
\keyword{internal}
