% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\name{orderly_backup}
\alias{orderly_backup}
\title{Create object for managing orderly db backups and do first backup}
\usage{
orderly_backup(config, backup_period)
}
\arguments{
\item{config}{Orderly config.}

\item{backup_period}{Period (in seconds) between DB backups.  This
is a guide only as backups happen on a API preroute - if more than
this many seconds have elapsed when the API receives a request then
a backup of the db will be performed.  This creates a copy of
orderly's destination database in\code{backup/db} with the same
filename as the destination database, even if that database typically
lives outside of the orderly tree.  In case of corruption of the
database, this backup can be manually moved into place.  This is only
needed if you are storing information alongside the core orderly
tables (as done by OrderlyWeb).}
}
\value{
orderly_backup_ object
}
\description{
Create object for managing orderly db backups and do first backup
}
\keyword{internal}
