% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup.R
\name{orderly_backup_}
\alias{orderly_backup_}
\title{Object for managing backing up orderly db}
\description{
Object for managing backing up orderly db

Object for managing backing up orderly db
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{config}}{Orderly config}

\item{\code{backup_period}}{How frequently should backup be run}

\item{\code{last_backup}}{Date-time when last backup was run}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{orderly_backup_$new()}}
\item \href{#method-check_backup}{\code{orderly_backup_$check_backup()}}
\item \href{#method-do_backup}{\code{orderly_backup_$do_backup()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create object, and do first backup.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{orderly_backup_$new(config, backup_period)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{Orderly config.}

\item{\code{backup_period}}{Period (in seconds) between DB backups.  This
is a guide only as backups happen on a API preroute - if more than
this many seconds have elapsed when the API receives a request then
a backup of the db will be performed.  This creates a copy of
orderly's destination database in\code{backup/db} with the same
filename as the destination database, even if that database typically
lives outside of the orderly tree.  In case of corruption of the
database, this backup can be manually moved into place.  This is only
needed if you are storing information alongside the core orderly
tables (as done by OrderlyWeb).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-check_backup"></a>}}
\if{latex}{\out{\hypertarget{method-check_backup}{}}}
\subsection{Method \code{check_backup()}}{
Check if a backup is due to be run, and run it if it is due.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{orderly_backup_$check_backup()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-do_backup"></a>}}
\if{latex}{\out{\hypertarget{method-do_backup}{}}}
\subsection{Method \code{do_backup()}}{
Backup the orderly database.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{orderly_backup_$do_backup()}\if{html}{\out{</div>}}
}

}
}
