% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runner.R
\name{orderly_runner}
\alias{orderly_runner}
\title{Orderly runner}
\usage{
orderly_runner(path, allow_ref = NULL, queue_id = NULL, workers = 1)
}
\arguments{
\item{path}{Path to use}

\item{allow_ref}{Allow git to change branches/ref for run.  If not
given, then we will look to see if the orderly configuration
disallows branch changes (based on the
\code{ORDERLY_API_SERVER_IDENTITY} environment variable and the
\code{master_only} setting of the relevant server block.}

\item{queue_id}{ID of an existing queue to connect to, creates a new one
if NULL}

\item{workers}{Number of workers to spawn}
}
\value{
A runner object, with methods designed for internal use only.
}
\description{
An orderly runner.  This is used to run reports as a server
process.  It's designed to be used in conjunction with OrderlyWeb,
so there is no "draft" stage and reports are committed as soon as
they are run.  This function is not intended for human end users,
only for creating automated tools for use with orderly.
}
\examples{
available <- redux::redis_available()
if (available) {
  path <- orderly:::prepare_orderly_git_example()
  runner <- orderly.server::orderly_runner(path[["local"]], workers = 0)
}
}
