% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{orderly_server_background}
\alias{orderly_server_background}
\title{Start a background orderly.server}
\usage{
orderly_server_background(path, port = 8321, log = NULL)
}
\arguments{
\item{path}{Path to an orderly root}

\item{port}{Port to listen on}

\item{log}{Path to log to}
}
\description{
Primarily for testing, start a server in the background
}
