% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Run orderly server}
\usage{
server(path, port, host = "0.0.0.0")
}
\arguments{
\item{path}{Path to serve}

\item{port}{Port to serve on}

\item{host}{Optional}
}
\description{
Run orderly server
}
