% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{workflow_missing_dependencies}
\alias{workflow_missing_dependencies}
\title{Get missing dependencies for list of reports.}
\usage{
workflow_missing_dependencies(path, reports)
}
\arguments{
\item{path}{Path to orderly root.}

\item{reports}{List of reports to get missing dependencies for,
this must be at a minimum a list of lists each item containing "name"
property of the report name. It can have additional data too but this
is ignored for checking dependencies.}
}
\value{
List with key report name and value is any missing dependencies.
}
\description{
Report A has a missing dependency if, report A has a "depends" block in
its `orderly.yml` which depends on latest version (or latest with some
search term) of report B and report B does not occur in the list of reports.
}
\keyword{internal}
