% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{workflow_summary}
\alias{workflow_summary}
\title{Get workflow summary for this set of reports}
\usage{
workflow_summary(path, reports, ref = NULL)
}
\arguments{
\item{path}{Path to orderly root.}

\item{reports}{List of reports to get workflow summary for,
this must be at a minimum a list of lists each item containing "name"
property of the report name. It can have additional data too but this
is ignored for checking dependencies.}
}
\value{
List containing reports in the order they will be
  queued including info about any dependencies between them. Also
  a list with key report name and value is any missing dependencies
  for that report.
}
\description{
This includes info on execution order of reports and any dependencies
between them and details of any missing dependencies.
}
\details{
Report A has a missing dependency if, report A has a "depends" block in
its `orderly.yml` which depends on latest version (or latest with some
search term) of report B and report B does not occur in the list of reports.
}
\keyword{internal}
