% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Run orderly server}
\usage{
server(path, port, host = "0.0.0.0", poll_interrupt = NULL,
  allow_ref = TRUE)
}
\arguments{
\item{path}{Path to serve}

\item{port}{Port to serve on}

\item{host}{Optional}

\item{poll_interrupt}{Interval (in ms) to poll for interrupt}

\item{allow_ref}{Allow git reference changing (passed through to
\code{orderly_runner}.}
}
\description{
Run orderly server
}
